task: changed t_raycast struct, changed raycaster algo, WIP, not tested

+++ raycaster/rotation.c
@@ -15,7 +15,7 @@
 void	rotation(int keycode, t_game game, t_raycast *ray)
 {
 	(void) game;
-	static double angle;
+	static float angle;
 	if (keycode == LEFT)
 	{
 		angle -= 0.1;
